refactor: move get_{ctf,challenge}_info to util
